refactor: remove dual layer of components context
Dual layer would mean that unused core
components would still get bundled.
Better to have a single file and expect
conflicts in it.
